==> Notebooks/1.1-InitialDataCleaning.ipynb <==
# Initial data cleaning

In this notebook we will clean and prepare the data for the analysis. We will perform the following steps:

1. Load the data
2. Eliminate irrelevant columns
3. Check for missing values
4. Handle missing values
5. Check unique values per column and filter categorical data
6. Save the clean data

# Importing libraries

import pandas as pd

# Load data and eliminate irrelevant columns

df_train = pd.read_csv("../Data/Raw/train.csv")
df_test = pd.read_csv("../Data/Raw/test.csv")
df_train = df_train.drop(columns="id")
df_test = df_test.drop(columns="id")

# Duplicate removal

df_train = df_train.drop_duplicates()
df_test = df_test.drop_duplicates()

# Checking missing values per column

missing = df_train.isna().mean() * 100
missing

### Missing values conclussion:

As seen in the missing values analysis, there are columns missing a lot of values. We will drop any column with more than 30% missing values. We will then drop all the rows with missing values as there is a lot of data and we can afford to lose some of it.

high_missing = df_train.columns[missing > 30]

df_train = df_train.drop(columns=high_missing)
df_test = df_test.drop(columns=high_missing)

df_train = df_train.dropna()
df_test = df_test.dropna()

df_train.info()

### Checking unique values per column

categorical = df_train.select_dtypes(include="object").columns.drop("class")

for column in categorical:
    unique = df_train[column].nunique()
    print(f"{column}: {unique}")

for column in categorical:
    print(f"\nRows per category in: \n{df_train[column].value_counts().head(10)}")

### Unique values conclussion:

Here it is clear that there are columns with a lot of unique values and most of that unique values only appear a couple of times. To reduce the amount of unique values we will unify all categories with less than 50000 appearances into a new category called 'Other'.

def unify_categories(df, column, threshold=50000):
    counts = df[column].value_counts()
    infrequent = counts[counts <= threshold].index
    df[column] = df[column].apply(lambda x: "Other" if x in infrequent else x)
    return df

for column in categorical:
    df_train = unify_categories(df_train, column)
    df_test = unify_categories(df_test, column)

# After unifying categories

for column in categorical:
    unique = df_train[column].nunique()
    print(f"{column}: {unique}")

for column in categorical:
    print(f"\nRows per category in: \n{df_train[column].value_counts().head(10)}")

# Duplicate removal

df_train = df_train.drop_duplicates()
df_test = df_test.drop_duplicates()

df_train.to_csv("../Data/PartiallyClean/train.csv", index=False)
df_test.to_csv("../Data/PartiallyClean/test.csv", index=False)
