==> Notebooks/1.2-EDA-FullDataCleaning.ipynb <==
# Exploratory data analysis and finalization of data cleaning process

All the figures will be saved at the [`Reports/Figures/`](https://github.com/faendal/MushroomEdibilityPrediction/tree/main/Reports/Figures) directory of the project.

The profiling report will be saved at the [`Reports/`](https://github.com/faendal/MushroomEdibilityPrediction/tree/main/Reports) directory of the project.
