==> Notebooks/1.3-DataTransformation.ipynb <==
# Data transformation and Sampling

Throughout this notebook, we will transform categorical data into numerical data and scale numerical data for the training set. We will use the following techniques:

- MinMaxScaler for numerical data
- LabelEncoder for class
- Dummy creation for categorical data

We will sample the data to work with a smaller dataset.

We will save the transformed data into [Data/WorkingData](https://github.com/faendal/MushroomEdibilityPrediction/tree/main/Data/WorkingData) folder.

Finally, we will save the transformation objects into [Models/Transformations](https://github.com/faendal/MushroomEdibilityPrediction/tree/main/Models/Transformations) folder.

import pickle
import pandas as pd
from sklearn.preprocessing import MinMaxScaler, LabelEncoder
from sklearn.model_selection import train_test_split

df = pd.read_csv('../Data/FullyClean/train.csv')

# Data type correction

types = {
    column: "category"
    for column in df.columns
    if df[column].dtype == "object"
}
df = df.astype(types)
df.info()

# Normalization

minmax = MinMaxScaler()

df[["stem-height", "stem-width"]] = minmax.fit_transform(df[["stem-height", "stem-width"]])

# Label encoding
labelEncoder = LabelEncoder()

df["class"] = labelEncoder.fit_transform(df["class"])

categorical = df.select_dtypes(include="category").columns

for column in categorical:
    unique = df[column].nunique()
    print(f"{column}: {unique}")

df = pd.get_dummies(df, columns=categorical, drop_first=False, dtype=int)

## Data Sampling

We currently have over 1.8 Million rows and 57 colums. This is a lot of data to work with. We will sample the data to reduce the size of the dataset and save it into the directory mentioned before

X = df.drop(columns=["class"])
y = df["class"]

# Sampling
X_sampled, _x, y_sampled, _y = train_test_split(X, y, test_size=0.7690006, stratify=y)

# DataFrame creation from sampled data
df_sampled = X_sampled.copy()
df_sampled["class"] = y_sampled
print(df_sampled.shape)

df_sampled.to_csv("../Data/WorkingData/train.csv", index=False)

filename = "../Models/Transformations/transformations.pkl"
variables = X.columns._values
pickle.dump([minmax, labelEncoder], open(filename, "wb"))
