==> Notebooks/2.2-HiperparameterOptimization.ipynb <==
## Hiperparameter optimization

After hipothesis testing, we will optimize the hyperparameters of:

1. Neural Network
2. XGBoost
3. Random Forest

We will use RandomizedSearch to determine some hyperparameters and Genetic Algorithm to determine the others.

import pickle
import seaborn as sns
from skopt import BayesSearchCV
from xgboost import XGBClassifier
from skopt.space import Real, Integer
from sklearn.neural_network import MLPClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import RandomizedSearchCV

sns.set_style("darkgrid")
sns.set_palette("Set1")

### Data loading

filename = "../Models/Parametrization/train.pkl"
X_train, y_train = pickle.load(open(filename, "rb"))

### RandomizedSearch

Here, we will make a first optimization of the hyperparameters of the models using RandomizedSearch. This will be a general approach and will help us to have a better idea of the hyperparameters that we should focus on later, when we use the Bayes Search Optimization.

mlp_param_grid = {
    "hidden_layer_sizes": [(25,), (50,), (25, 25), (50, 50)],
    "activation": ["relu", "tanh"],
    "solver": ["adam", "sgd"],
    "alpha": [0.001, 0.01],
    "learning_rate": ["adaptive"],
    "max_iter": [500, 1000],
}

xgb_param_grid = {
    "n_estimators": [50, 100, 200],
    "learning_rate": [0.01, 0.1, 0.2],
    "max_depth": [3, 5, 7],
    "subsample": [0.7, 0.8, 0.9],
    "colsample_bytree": [0.7, 0.8, 0.9],
}

rf_param_grid = {
    "n_estimators": [50, 100, 200],
    "max_depth": [5, 10, 20],
    "min_samples_split": [3, 5, 7],
    "min_samples_leaf": [2],
    "max_features": ["auto", "sqrt", "log2"],
}

mlp = MLPClassifier(random_state=42)
mlp_grid_search = RandomizedSearchCV(
    mlp, mlp_param_grid, cv=3, scoring="accuracy", verbose=2, n_jobs=-1
)
mlp_grid_search.fit(X_train, y_train)
print("Best parameters for MLPClassifier:", mlp_grid_search.best_params_)

xgb = XGBClassifier(random_state=42)
xgb_grid_search = RandomizedSearchCV(
    xgb, xgb_param_grid, cv=3, scoring="accuracy", verbose=2, n_jobs=-1
)
xgb_grid_search.fit(X_train, y_train)
print("Best parameters for XGBoostClassifier:", xgb_grid_search.best_params_)

rf = RandomForestClassifier(random_state=42)
rf_grid_search = RandomizedSearchCV(
    rf, rf_param_grid, cv=3, scoring="accuracy", verbose=2, n_jobs=-1
)
rf_grid_search.fit(X_train, y_train)
print("Best parameters for RandomForest:", rf_grid_search.best_params_)

### Bayes Search Optimization

Here we will make a more throrough optimization of the hyperparameters of the models using the Bayes Search Optimization. This will be a more specific approach and will take into account the hyperparameters that we found to be more important in the RandomizedSearch.

# adam, adaptive, relu

mlp_param_grid = {
    "alpha": Real(0.001, 0.1),
    "max_iter": Integer(500, 1000),
    "momentum": Real(0.2, 0.4),
    "learning_rate_init": Real(0.1, 0.3)
}

xgb_param_grid = {
    "n_estimators": Integer(50, 200),
    "learning_rate": Real(0.01, 0.2),
    "max_depth": Integer(3, 7),
    "subsample": Real(0.7, 0.9),
    "colsample_bytree": Real(0.7, 0.9),
}

# log2
rf_param_grid = {
    "n_estimators": Integer(50, 200),
    "max_depth": Integer(5, 20),
    "min_samples_split": Integer(3, 7),
}

mlp = MLPClassifier(random_state=42, activation="relu", solver="adam", learning_rate="adaptive", hidden_layer_sizes=(50,50))
mlp_bayes = BayesSearchCV(
    estimator=mlp,
    search_spaces=mlp_param_grid,
    cv=5,
    scoring="accuracy",
    verbose=2,
    n_iter=50,
    n_jobs=-1,
    refit=True,
)
mlp_bayes.fit(X_train, y_train)
print("Best parameters for MLPClassifier:", mlp_bayes.best_params_)

xgb = XGBClassifier(random_state=42)
xgb_bayes = BayesSearchCV(
    estimator=xgb,
    search_spaces=xgb_param_grid,
    cv=5,
    scoring="accuracy",
    verbose=2,
    n_iter=50,
    n_jobs=-1,
    refit=True,
)
xgb_bayes.fit(X_train, y_train)
print("Best parameters for XGBoostClassifier:", xgb_bayes.best_params_)

rf = RandomForestClassifier(random_state=42, max_features="log2", min_samples_leaf=2)
rf_bayes = BayesSearchCV(
    estimator=rf,
    search_spaces=rf_param_grid,
    cv=5,
    scoring="accuracy",
    verbose=2,
    n_iter=50,
    n_jobs=-1,
    refit=True,
)
rf_bayes.fit(X_train, y_train)
print("Best parameters for RandomForest:", rf_bayes.best_params_)

filename = "../Models/ML/Hyperparametrized/hyper_models.pkl"
with open(filename, "wb") as file:
    pickle.dump(
        [
            mlp_grid_search,
            xgb_grid_search,
            rf_grid_search,
            mlp_bayes,
            xgb_bayes,
            rf_bayes,
        ],
        file,
    )
