==> Notebooks/2.3-ModelSelection.ipynb <==
# Model Selection

In this notebook, we will train models with the parameters we got from hiperparametrization, compare the performance of said models and select and train the model that will be used for deployment

import pickle
import pandas as pd
from sklearn import metrics
from xgboost import XGBClassifier
from sklearn.neural_network import MLPClassifier
from sklearn.ensemble import RandomForestClassifier

filename = "../Models/ML/Hyperparametrized/hyper_models.pkl"
with open(filename, "rb") as file:
    mlp_grid, xgb_grid, rf_grid, mlp_bayes, xgb_bayes, rf_bayes = pickle.load(file)

filename = "../Models/Parametrization/train.pkl"
X_train, y_train = pickle.load(open(filename, "rb"))
filename = "../Models/Parametrization/test.pkl"
X_test, y_test = pickle.load(open(filename, "rb"))

filename = "../Models/Transformations/transformations.pkl"
minmax, labelEncoder = pickle.load(open(filename, "rb"))

hyper_models = pd.DataFrame(index=["Acuracy", "ROC AUC", "F1 Score", "Recall Poisonous"])

### MLPClassifier

mlp_grid.best_params_

mlp_grid_model = MLPClassifier(
    activation="relu",
    hidden_layer_sizes=(50, 50),
    solver="adam",
    learning_rate="adaptive",
    alpha=0.01,
    max_iter=500,
)

mlp_grid_model.fit(X_train, y_train)

Y_pred = mlp_grid_model.predict(X_test)
print(
    metrics.classification_report(
        y_true=y_test, y_pred=Y_pred, target_names=labelEncoder.classes_
    )
)

hyper_models["NN Grid"] = [
    metrics.accuracy_score(y_test, Y_pred),
    metrics.roc_auc_score(y_test, Y_pred),
    metrics.f1_score(y_test, Y_pred),
    metrics.recall_score(y_test, Y_pred, pos_label=1),
]

mlp_bayes.best_params_

mlp_bayes_model = MLPClassifier(
    activation="relu",
    hidden_layer_sizes=(50, 50),
    solver="adam",
    learning_rate="adaptive",
    alpha=0.001,
    learning_rate_init=0.1,
    momentum=0.4,
    max_iter=909,
)

mlp_bayes_model.fit(X_train, y_train)

Y_pred = mlp_bayes_model.predict(X_test)
print(
    metrics.classification_report(
        y_true=y_test, y_pred=Y_pred, target_names=labelEncoder.classes_
    )
)

hyper_models["NN Bayes"] = [
    metrics.accuracy_score(y_test, Y_pred),
    metrics.roc_auc_score(y_test, Y_pred),
    metrics.f1_score(y_test, Y_pred),
    metrics.recall_score(y_test, Y_pred, pos_label=1),
]

### XGBoost

xgb_grid.best_params_

xgb_grid_model = XGBClassifier(
    n_estimators=100,
    max_depth=7,
    learning_rate=0.2,
    subsample=0.8,
    colsample_bytree=0.7,
    use_label_encoder=False,
    eval_metric="logloss",
    random_state=42,
)
xgb_grid_model.fit(X_train, y_train)

Y_pred = xgb_grid_model.predict(X_test)
Y_pred_proba = xgb_grid_model.predict_proba(X_test)[:, 1]

print(
    metrics.classification_report(
        y_true=y_test, y_pred=Y_pred, target_names=labelEncoder.classes_
    )
)

hyper_models["XGBoost Grid"] = [
    metrics.accuracy_score(y_test, Y_pred),
    metrics.roc_auc_score(y_test, Y_pred_proba),
    metrics.f1_score(y_test, Y_pred),
    metrics.recall_score(y_test, Y_pred, pos_label=1),
]

xgb_bayes.best_params_

xgb_bayes_model = XGBClassifier(
    n_estimators=200,
    max_depth=7,
    learning_rate=0.2,
    subsample=0.7,
    colsample_bytree=0.9,
    use_label_encoder=False,
    eval_metric="logloss",
    random_state=42,
)
xgb_bayes_model.fit(X_train, y_train)

Y_pred = xgb_bayes_model.predict(X_test)
Y_pred_proba = xgb_bayes_model.predict_proba(X_test)[:, 1]

print(
    metrics.classification_report(
        y_true=y_test, y_pred=Y_pred, target_names=labelEncoder.classes_
    )
)

hyper_models["XGBoost Bayes"] = [
    metrics.accuracy_score(y_test, Y_pred),
    metrics.roc_auc_score(y_test, Y_pred_proba),
    metrics.f1_score(y_test, Y_pred),
    metrics.recall_score(y_test, Y_pred, pos_label=1),
]

### RandomForest

rf_grid.best_params_

rf_grid_model = RandomForestClassifier(
    n_estimators=100,
    max_depth=10,
    min_samples_split=5,
    min_samples_leaf=2,
    random_state=42,
    max_features="log2",
    n_jobs=-1,
)
rf_grid_model.fit(X_train, y_train)

Y_pred = rf_grid_model.predict(X_test)
Y_pred_proba = rf_grid_model.predict_proba(X_test)[:, 1]

print(
    metrics.classification_report(
        y_true=y_test, y_pred=Y_pred, target_names=labelEncoder.classes_
    )
)

hyper_models["Random Forest Grid"] = [
    metrics.accuracy_score(y_test, Y_pred),
    metrics.roc_auc_score(y_test, Y_pred_proba),
    metrics.f1_score(y_test, Y_pred),
    metrics.recall_score(y_test, Y_pred, pos_label=1),
]

rf_bayes.best_params_

rf_bayes_model = RandomForestClassifier(
    n_estimators=139,
    max_depth=20,
    min_samples_split=3,
    min_samples_leaf=2,
    random_state=42,
    max_features="log2",
    n_jobs=-1,
)
rf_bayes_model.fit(X_train, y_train)

Y_pred = rf_bayes_model.predict(X_test)
Y_pred_proba = rf_bayes_model.predict_proba(X_test)[:, 1]

print(
    metrics.classification_report(
        y_true=y_test, y_pred=Y_pred, target_names=labelEncoder.classes_
    )
)

hyper_models["Random Forest Bayes"] = [
    metrics.accuracy_score(y_test, Y_pred),
    metrics.roc_auc_score(y_test, Y_pred_proba),
    metrics.f1_score(y_test, Y_pred),
    metrics.recall_score(y_test, Y_pred, pos_label=1),
]

## Evaluation and model selection

hyper_models

## Conclusion

After training the models and evaluating them, we get 4 models with high performance:

1. Neural Network Grid
2. XGBoost Grid
3. XGBoost Bayes
4. RandomForest Bayes

In the end, we selected XGBoost Bayes, as it is the model with the best performance in all of the metrics we evaluated and is one of the fastest to train.

We will now train it with the entire dataset and save it for deployment.

X = pd.concat([X_train, X_test])
Y = pd.concat([y_train, y_test])

xgb_bayes_model.fit(X, Y)

filename = "../Models/ML/Deployment/xgb_bayes.pkl"
with open(filename, "wb") as file:
    pickle.dump([xgb_bayes_model], file)
