==> Notebooks/2.4-Pipeline.ipynb <==
import pickle
import pandas as pd
from sklearn.pipeline import Pipeline
from PipelineClasses import (
    CorrectTypesTransformer,
    ExtremeOutlierCleaningTransformer,
    PretrainedMinMaxScaler,
    PretrainedLabelEncoder,
    DummyTransformer,
    NumericalImputer,
    CategoricalImputer,
)

filename = "../Models/ML/Deployment/xgb_bayes.pkl"
with open(filename, 'rb') as file:
    model = pickle.load(file)
xgb = model[0]

filename = "../Models/Transformations/transformations.pkl"
with open(filename, 'rb') as file:
    minmax, labelEncoder = pickle.load(file)

pipeline = Pipeline(
    [
        ("correct_types", CorrectTypesTransformer()),
        ("numerical_imputer", NumericalImputer(strategy="mean")),
        ("categorical_imputer", CategoricalImputer(strategy="most_frequent")),
        ("outlier_cleaning", ExtremeOutlierCleaningTransformer()),
        ("dummy_encoder", DummyTransformer()),
        ("scaler", PretrainedMinMaxScaler(scaler=minmax)),
        ("model", xgb),
    ]
)
pipeline

filename = "../Models/Pipe/pipeline.pkl"
with open(filename, 'wb') as file:
    pickle.dump(pipeline, file)
