==> Notebooks/3-ModelDeployment.ipynb <==
import pickle
import pandas as pd
import urllib.request
import streamlit as st

%%cmd
streamlit run --server.port 8501 app.py 

# %%bash
# nohup streamlit run app.py &

# %%bash 
# npm install localtunnel

# print(
#     "Password/Enpoint IP for localtunnel is:",
#     urllib.request.urlopen("https://ipv4.icanhazip.com")
#     .read()
#     .decode("utf8")
#     .strip("\n"),
# )

#%%bash
#npx localtunnel --port 8501
